Listing List fix
Changed listing list to display an incrementing number on the buttons, as opposed to the rent value. This was done as I was unable to get the list to change the rent value that was previously displayed on the buttons, after a listing had been updated. For some reason this morning I also had to change a number of package declarations and reinstall some packages.